add home page content that's only shown to authenticated users. Conditionally display that content in the home page when authenticated
